refactor: global/rq 패키지 생서하여 rq 관련 코드 이동
