+++ backend/src/main/java/com/NBE_4_5_2/Team5/domain/user/user/controller/UserController.kt
@@ -5,7 +5,7 @@ import com.NBE_4_5_2.Team5.domain.user.user.dto.UserDto
 import com.NBE_4_5_2.Team5.domain.user.user.dto.UserUpdateRequest
 import com.NBE_4_5_2.Team5.domain.user.user.service.UserAuthService
 import com.NBE_4_5_2.Team5.domain.user.user.service.UserService
-import com.NBE_4_5_2.Team5.global.Rq
+import com.NBE_4_5_2.Team5.global.rq.Rq
 import com.NBE_4_5_2.Team5.global.dto.Empty
 import com.NBE_4_5_2.Team5.global.response.RsData
 import io.swagger.v3.oas.annotations.Operation
