+++ backend/src/main/java/com/NBE_4_5_2/Team5/domain/user/user/service/UserService.kt
@@ -9,7 +9,7 @@ import com.NBE_4_5_2.Team5.domain.user.user.entity.Role
 import com.NBE_4_5_2.Team5.domain.user.user.entity.User
 import com.NBE_4_5_2.Team5.domain.user.user.repository.UserRepository
 import com.NBE_4_5_2.Team5.domain.user.user.service.email.EmailService
-import com.NBE_4_5_2.Team5.global.Rq
+import com.NBE_4_5_2.Team5.global.rq.Rq
 import com.NBE_4_5_2.Team5.global.exception.ServiceException
 import com.NBE_4_5_2.Team5.global.exception.security.AuthenticationNotValidException
 import com.NBE_4_5_2.Team5.global.exception.security.TokenNotFoundException
