+++ backend/src/main/java/com/NBE_4_5_2/Team5/global/rq/DevRq.kt
@@ -1,4 +1,4 @@
-package com.NBE_4_5_2.Team5.global
+package com.NBE_4_5_2.Team5.global.rq
 
 import com.NBE_4_5_2.Team5.domain.user.user.service.UserService
 import jakarta.servlet.http.Cookie
@@ -7,7 +7,7 @@ import jakarta.servlet.http.HttpServletResponse
 import org.springframework.context.annotation.Profile
 import org.springframework.stereotype.Component
 import org.springframework.web.context.annotation.RequestScope
-import java.util.Optional
+import java.util.*
 
 @Component
 @RequestScope
