+++ backend/src/main/java/com/NBE_4_5_2/Team5/global/rq/Rq.kt
@@ -1,4 +1,4 @@
-package com.NBE_4_5_2.Team5.global
+package com.NBE_4_5_2.Team5.global.rq
 
 import java.util.Optional
