+++ backend/src/main/java/com/NBE_4_5_2/Team5/global/security/CustomAuthenticationFilter.kt
@@ -4,7 +4,7 @@ import com.NBE_4_5_2.Team5.domain.user.user.dto.AuthToken
 import com.NBE_4_5_2.Team5.domain.user.user.entity.User
 import com.NBE_4_5_2.Team5.domain.user.user.service.UserAuthService
 import com.NBE_4_5_2.Team5.domain.user.user.service.UserService
-import com.NBE_4_5_2.Team5.global.Rq
+import com.NBE_4_5_2.Team5.global.rq.Rq
 import jakarta.servlet.FilterChain
 import jakarta.servlet.ServletException
 import jakarta.servlet.http.HttpServletRequest
